Refactor upload presentation layer
